requirements.txt
# AI 价值投资系统依赖
# 第一阶段（P0）：数据层自动化

# 核心依赖
tushare==1.2.89  # A 股财务/行情数据
akshare==1.12.0  # 备用数据源
pandas==2.1.0  # 数据处理
numpy==1.24.0  # 数值计算
pyyaml==6.0.1  # 配置文件

# OCR（第三阶段 P2）
paddlepaddle==2.5.0  # PaddleOCR
paddleocr==2.7.0

# 调度（第四阶段 P3）
celery==5.3.0
redis==4.6.0

# 通知（第二阶段 P1）
requests==2.31.0

# 测试
pytest==7.4.0
